start eto refactoring
