+++ app/lib/api/eto/EtoApi.interfaces.ts
@@ -1,4 +1,4 @@
-import { DeepPartial } from "../../../types";
+import { DeepPartial, DeepReadonlyObject } from "../../../types";
 import * as YupTS from "../../yup-ts";
 import { TEtoDocumentTemplates } from "./EtoFileApi.interfaces";
 
@@ -251,10 +251,10 @@ export type TGeneralEtoData = {
 };
 
 // this is comming from the /etos endpoint for investors dashboard
-export type TInvestorEtoData = TEtoSpecsData & {
+export type TPublicEtoData = DeepReadonlyObject<TEtoSpecsData & {
   company: TCompanyEtoData;
   etoId: string;
-};
+}>;
 
 export const GeneralEtoDataType = YupTS.object({
   ...EtoTermsType.shape,
