+++ app/modules/actions.ts
@@ -1,11 +1,9 @@
 import { TDictionaryValues } from "../types";
 import { accessWalletActions } from "./accessWallet/actions";
 import { authActions } from "./auth/actions";
-import { dashboardActions } from "./dashboard/actions";
 import { depositModalActions } from "./depositEthModal/actions";
 import { etoDocumentsActions } from "./eto-documents/actions";
 import { etoFlowActions } from "./eto-flow/actions";
-import { etoActions } from "./eto/actions";
 import { gasActions } from "./gas/actions";
 import { genericModalActions } from "./genericModal/actions";
 import { icbmWalletBalanceModalActions } from "./icbmWalletBalanceModal/actions";
@@ -15,6 +13,7 @@ import { investmentFlowActions } from "./investmentFlow/actions";
 import { kycActions } from "./kyc/actions";
 import { notificationActions } from "./notifications/actions";
 import { personProfileModalActions } from "./personProfileModal/actions";
+import { etoActions } from "./public-etos/actions";
 import { routingActions } from "./routing/actions";
 import { settingsActions } from "./settings/actions";
 import { formSingleFileUploadActions } from "./shared/formSingleFileUpload/actions";
@@ -31,7 +30,6 @@ import { web3Actions } from "./web3/actions";
 
 /** You should add new actions also here (with a namespace).*/
 export const actions = {
-  dashboard: dashboardActions,
   settings: settingsActions,
   immutableStorage: immutableStorageActions,
   verifyEmail: verifyEmailActions,
@@ -49,7 +47,7 @@ export const actions = {
   notifications: notificationActions,
   etoFlow: etoFlowActions,
   etoDocuments: etoDocumentsActions,
-  eto: etoActions,
+  publicEtos: etoActions,
   formSingleFileUpload: formSingleFileUploadActions,
   remoteFile: remoteFileActions,
   depositEthModal: depositModalActions,
@@ -66,7 +64,6 @@ export const actions = {
  * Merge all actions into main action object. Add new set of actions here.
  */
 export const allActions = {
-  ...dashboardActions,
   ...tokenPriceActions,
   ...etoDocumentsActions,
   ...verifyEmailActions,
