+++ app/modules/investmentFlow/reducer.ts
@@ -1,6 +1,6 @@
 import BigNumber from "bignumber.js";
 
-import { TInvestorEtoData } from "../../lib/api/eto/EtoApi.interfaces";
+import { TPublicEtoData } from "../../lib/api/eto/EtoApi.interfaces";
 import { AppReducer } from "../../store";
 import { DeepReadonly } from "../../types";
 import { multiplyBigNumbers } from "../../utils/BigNumberUtils";
@@ -30,22 +30,11 @@ export enum EInvestmentErrorState {
   NotEnoughEtherForGas = "not_enough_ether_for_gas",
 }
 
-export interface ICalculatedContribution {
-  isWhitelisted: boolean;
-  minTicketEurUlps: BigNumber;
-  maxTicketEurUlps: BigNumber;
-  equityTokenInt: BigNumber;
-  neuRewardUlps: BigNumber;
-  maxCapExceeded: boolean;
-}
-
 export interface IInvestmentFlowState {
   euroValueUlps: string;
   ethValueUlps: string;
   investmentType: EInvestmentType;
-  eto?: TInvestorEtoData;
   errorState?: EInvestmentErrorState;
-  calculatedContribution?: ICalculatedContribution;
   gasAmount: string;
   gasPrice: string;
 }
@@ -73,7 +62,6 @@ export const investmentFlowReducer: AppReducer<IInvestmentFlowState> = (
         investmentType: action.payload.type,
         euroValueUlps: "",
         errorState: undefined,
-        calculatedContribution: undefined,
       };
     case "INVESTMENT_FLOW_SET_GAS_PRICE":
       return {
@@ -83,7 +71,6 @@ export const investmentFlowReducer: AppReducer<IInvestmentFlowState> = (
     case "INVESTMENT_FLOW_SET_ETO":
       return {
         ...state,
-        eto: action.payload.eto,
       };
     case "INVESTMENT_FLOW_SET_INVESTMENT_ERROR_STATE":
       return {
@@ -99,11 +86,6 @@ export const investmentFlowReducer: AppReducer<IInvestmentFlowState> = (
       return {
         ...state,
         euroValueUlps: action.payload.value,
-      };
-    case "INVESTMENT_FLOW_SET_CALCULATED_CONTRIBUTION":
-      return {
-        ...state,
-        calculatedContribution: action.payload.contrib,
       };
   }
